fix(api): reconcile ContextLister status reads with main's scoped-store mitigation
main merged ga-cdmx6x (5aedfbe38) since this branch's last rebase, adding a
throwaway ctx-bound BdStore mechanism (cmd/gc/scoped_store.go +
State.ScopedStoreLike) at the same 3 status/store-health call sites this
branch's ContextLister capability (ga-oeeggk/ga-yxwid1) now covers directly
via the store's own long-lived runner. Per ga-f1frwf's Decision B-supersedes-
Decision-A design, rebase onto main and remove the scoped-store mechanism
entirely rather than running both side by side: scoped_store.go/_test.go,
handler_status_scoped_store_test.go, State.ScopedStoreLike, the bd_env.go
NoRecovery env variants it alone used, and BdStore.Dir() (its only caller).
loadStatusSessionSnapshot and statusListStoreWithTimeout revert to their
pre-existing signatures now that no cityPath/cfg/state plumbing is needed.
